LayoutTests/imported/w3c/web-platform-tests/mediacapture-streams/MediaStream-MediaElement-srcObject.https-expected.txt

When prompted, accept to share your video stream.

Description

This test checks that the MediaStream object returned by the success callback in getUserMedia can be properly assigned to a video element via the srcObject attribute.



PASS Tests that a MediaStream can be assigned to a video element with srcObject 
PASS Tests that a MediaStream assigned to a video element is not seekable 
PASS Tests that a MediaStream assigned to a video element is in readyState HAVE_NOTHING initially 
PASS Tests that a MediaStream assigned to a video element has expected duration 
PASS Tests that a video element with a MediaStream assigned is not preloaded 
PASS Tests that a video element with a MediaStream assigned ignores playbackRate attributes (defaultPlaybackRate is identical) 
PASS Tests that a video element with a MediaStream assigned ignores playbackRate attributes (defaultPlaybackRate is different) 
FAIL Tests that a media element with an assigned MediaStream reports the played attribute as expected assert_equals: A MediaStream's end MUST return the last known currentTime expected 0.252313446 but got 0.252291624
FAIL Tests that a media element with an assigned MediaStream reports the currentTime attribute as expected assert_equals: The UA MUST ignore attempts to set the currentTime attribute (restart) expected 0.250999782 but got 0.251026206
PASS Tests that a media element with an assigned MediaStream starts its timeline at 0 regardless of when the MediaStream was created 
FAIL Tests that a media element with an assigned MediaStream does not advance currentTime while paused assert_between_exclusive: currentTime does not skip ahead after pause expected a number greater than 0 and less than 0.5 but got 0.755360896
PASS Tests that the loop attribute has no effect on a media element with an assigned MediaStream 
PASS Tests that a media element with an assigned MediaStream ends when the MediaStream becomes inactive through tracks ending 
FAIL Tests that an audio element with an assigned MediaStream ends when the MediaStream becomes inaudible through audio tracks ending assert_true: HTMLAudioElement becomes ended asynchronously when its MediaStream provider becomes inaudible expected true got false
PASS Tests that a media element with an assigned MediaStream ends when the MediaStream becomes inactive through track removal 
FAIL Tests that an audio element with an assigned MediaStream ends when the MediaStream becomes inaudible through track removal assert_true: HTMLAudioElement becomes ended asynchronously when its MediaStream provider becomes inaudible expected true got false
